// projects/09-translator/src/types.d.ts
import { AUTO_LANGUAGE, SUPPORTED_LANGUAGES } from "./consts"

export interface State {
  fromLanguage: FromLanguage
  toLanguage: Language
  originalText: string
  translatedText: string
  loading: boolean
}

export type Action =
  { type: 'SWAP_LANGUAGES' } |
  { type: 'SET_FROM_LANGUAGE', payload: FromLanguage } |
  { type: 'SET_TO_LANGUAGE', payload: Language } |
  { type: 'SET_ORIGINAL_TEXT', payload: string } |
  { type: 'SET_TRANSLATED_TEXT', payload: string }

export type Language = keyof typeof SUPPORTED_LANGUAGES
export type AutoLanguage = typeof AUTO_LANGUAGE
export type FromLanguage = Language | AutoLanguage

export enum SectionType {
  From = 'from',
  To = 'to'
}